Worked on the Data Handler!
Data handler changed to work with the Orientation Event listener and
calculate the R1 and R2 values. Some testing has to be done with the
phone given that the axis change between the Surface and phone devices.
However, the code now gives indication to the User if they are inside
the range to start the test and can handle motion.